Moved shuffleArray from separate module to within defineGoals. #28
