--- client/utils/shuffle-array.js
@@ -1,12 +1,0 @@
-function shuffleArray(array) {
-  let shuffled = array.slice()
-  for (let i = shuffled.length - 1; i > 0; i--) {
-    let j = Math.floor(Math.random() * (i + 1))
-    let temp = shuffled[i]
-    shuffled[i] = shuffled[j]
-    shuffled[j] = temp
-  }
-  return shuffled
-}
-
-module.exports = shuffleArray
